// front-exam/src/app/apps/common/model/clientDTO.ts
/**
 * Exam Management API
 * This API exposes endpoints to manage client.
 *
 * OpenAPI spec version: 1.0
 * Contact: yessinelleuch00@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { CompteDTO } from './compteDTO';

export interface ClientDTO { 
    id?: string;
    cin?: string;
    firstName?: string;
    lastName?: string;
    comptes?: Array<CompteDTO>;
}